+++ src/main/java/com/example/order/domain/type/RiderTransportation.java
@@ -1,0 +1,6 @@
+package com.example.order.domain.type;
+
+public enum RiderTransportation {
+    WALK, BICYCLE, MOTORBIKE, CAR
+
+}
